Build: check formatting with a pre-commit git hook
I wanted this to be an auto-format files pre-commit hook. Unfortunately, the documented way by Prettier for formatting staged files and re-adding them before making the actual commit has a really big flaw: if you add files to git with `git add --patch`, then **the full file** will be added to the commit. This is a no-go for me.